Refactor k-form
